backend/requirements.txt
# RivalEdge Backend Requirements
# Python 3.11+

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.0

# Auth
PyJWT[crypto]==2.10.1
cryptography==43.0.3

# Database
supabase==2.10.0

# HTTP client
httpx==0.28.0

# Validation
pydantic[email]==2.10.1

# AI
anthropic==0.40.0

# Scraping
playwright==1.49.0

# Email
resend==2.4.0
markdown==3.7

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0

# Dev
python-dotenv==1.0.1
